scripts/requirements.txt
numpy==1.16.0
Pillow==5.1.1
